Major update to how the game addresses ships roles, now roles are defined in files in the Content/ShipRoles folder, ShipUpkeep.xml has been scrapped to prevent conflicts
--HG--
branch : McShooterz2

+++ Ship_Game/SavedGame.cs
@@ -845,7 +845,7 @@ namespace Ship_Game
 
 			public float InCombatTimer;
 
-			public int experience;
+			public float experience;
 
 			public int kills;
